backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
Werkzeug==2.3.7
pyodbc==5.0.1
python-dotenv==1.0.0
